software/phase2/requirements.txt
cmake
opencv-python
numpy
apriltag
cmapy
smbus
crcmod
matplotlib
pyserial
imutils
mediapipe
